+++ src/renderer/render/media.test.ts
@@ -1,66 +1,62 @@
 import { media } from './media';
+import { parse } from '../../parser';
 import { html } from 'typed-dom';
 
 describe('Unit: renderer/render/media', () => {
 
-  function image(url: string) {
-    return html('img', { 'data-src': url });
-  }
-
   describe('media', () => {
     it('twitter', done => {
-      media(image('http://twitter.com/hourenso_u/status/856828123882676225'), {
+      media(parse('!http://twitter.com/hourenso_u/status/856828123882676225').querySelector('img')!, {
         twitter: () => void done(false) || html('div')
       });
-      media(image('https://twitter.com/hourenso_u/status/856828123882676225'), {
+      media(parse('!https://twitter.com/hourenso_u/status/856828123882676225').querySelector('img')!, {
         twitter: () => void done() || html('div')
       });
     });
 
     it('youtube', done => {
-      media(image('http://youtu.be/xRF7WIZV4lA'), {
+      media(parse('!http://youtu.be/xRF7WIZV4lA').querySelector('img')!, {
         youtube: () => void done(false) || html('div')
       });
-      media(image('https://youtu.be/xRF7WIZV4lA'), {
+      media(parse('!https://youtu.be/xRF7WIZV4lA').querySelector('img')!, {
         youtube: () => void done() || html('div')
       });
-      media(image('http://www.youtube.com/watch?v=xRF7WIZV4lA'), {
+      media(parse('!http://www.youtube.com/watch?v=xRF7WIZV4lA').querySelector('img')!, {
         youtube: () => void done(false) || html('div')
       });
-      media(image('https://www.youtube.com/watch?v=xRF7WIZV4lA'), {
+      media(parse('!https://www.youtube.com/watch?v=xRF7WIZV4lA').querySelector('img')!, {
         youtube: () => void done() || html('div')
       });
     });
 
     it('gist', done => {
-      media(image('http://gist.github.com/falsandtru/cdf4a19b70012b0d6e3c9e1ec021e557'), {
+      media(parse('!http://gist.github.com/falsandtru/cdf4a19b70012b0d6e3c9e1ec021e557').querySelector('img')!, {
         gist: () => void done(false) || html('div')
       });
-      media(image('https://gist.github.com/falsandtru/cdf4a19b70012b0d6e3c9e1ec021e557'), {
+      media(parse('!https://gist.github.com/falsandtru/cdf4a19b70012b0d6e3c9e1ec021e557').querySelector('img')!, {
         gist: () => void done() || html('div')
       });
     });
 
     it('slideshare', done => {
-      media(image('http://www.slideshare.net/Slideshare/get-started-with-slide-share'), {
+      media(parse('!http://www.slideshare.net/Slideshare/get-started-with-slide-share').querySelector('img')!, {
         slideshare: () => void done(false) || html('div')
       });
-      media(image('https://www.slideshare.net/Slideshare/get-started-with-slide-share'), {
+      media(parse('!https://www.slideshare.net/Slideshare/get-started-with-slide-share').querySelector('img')!, {
         slideshare: () => void done() || html('div')
       });
     });
 
     it('pdf', done => {
-      media(image('http://example.pdf'), {
+      media(parse('!http://example.pdf').querySelector('img')!, {
         pdf: () => void done(false) || html('div')
       });
-      assert(media(image('http://example.com/example.pdf'), {})!.querySelector('object')!.getAttribute('type') === 'application/pdf');
+      assert(media(parse('!http://example.com/example.pdf').querySelector('img')!, {})!.querySelector('object')!.getAttribute('type') === 'application/pdf');
       done();
     });
 
     it('image', () => {
-      assert(media(image('/'), {})!.getAttribute('src') === new URL('/', location.href).href);
-      assert(media(image('/'), {})!.getAttribute('alt') === '');
+      assert(media(parse('!{/}').querySelector('img')!, {})!.matches(`[src="${new URL('/', location.href).href}"][alt=""]`));
     });
 
   });