+++ src/renderer/render/media/gist.ts
@@ -11,7 +11,7 @@ export function gist(url: URL): HTMLElement | undefined {
   if (!origins.has(url.origin)) return;
   if (!url.pathname.match(/^\/[\w-]+?\/\w{32}(?!\w)/)) return;
   if (cache.has(url.href)) return cache.get(url.href)!.cloneNode(true);
-  return HTML.div({ style: 'position: relative;' }, [HTML.em(`loading ${url.href}`)], (f, tag) => {
+  return HTML.div({ class: 'media', style: 'position: relative;' }, [HTML.em(`loading ${url.href}`)], (f, tag) => {
     const outer = f(tag);
     void $.ajax(`${url.href}.json`, {
       dataType: 'jsonp',