+++ src/renderer/render/media/twitter.ts
@@ -27,7 +27,7 @@ export function twitter(url: URL): HTMLElement | undefined {
     window.twttr && void window.twttr.widgets.load(el);
     return el;
   }
-  return HTML.div({ style: 'position: relative;' }, [HTML.em(`loading ${url.href}`)], (f, tag) => {
+  return HTML.div({ class: 'media', style: 'position: relative;' }, [HTML.em(`loading ${url.href}`)], (f, tag) => {
     const outer = f(tag);
     void $.ajax(`https://publish.twitter.com/oembed?url=${url.href.replace('?', '&')}&omit_script=true`, {
       dataType: 'jsonp',